attempting relative URL
Although the page name and page render under "extras" I am trying "Go to the [“old Run Selector” instructions](../_extras/old-ncbi/) page" to see if a relative address can work. This is because the "old-ncbi" page used a datacarpentry URL and that page doesn't exist. NOT SURE WHY IT WORKS FOR EXTRAS ON FILE MENU.